[webview_flutter_lwe] Own the cookie manager channel at the plugin level
The cookie manager channel was registered in the WebView constructor with a raw pointer to that view and was never unregistered. Calling the cookie manager before any WebView was created raised a MissingPluginException, and calling it after the last view was disposed dispatched through a destroyed WebView.

The LWE cookie manager is a process-wide singleton, so the channel is now created and owned by the plugin for its whole lifetime and the handler no longer touches a WebView instance.
